backend/requirements.txt
# FastAPI and server
fastapi>=0.110.0,<0.111.0
uvicorn[standard]>=0.27.0,<0.28.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23,<2.1.0
psycopg[binary]>=3.1.13,<3.2.0
alembic>=1.12.1,<1.13.0

# Redis
redis>=5.0.1,<5.1.0

# Payment processing
stripe>=7.0.0,<8.0.0
paypalrestsdk>=1.13.0,<2.0.0

# Email
emails>=0.6.0,<0.7.0

# Utilities
pydantic>=2.5.0,<2.6.0
pydantic-settings>=2.1.0,<2.2.0
python-dotenv>=1.0.0,<1.1.0
httpx>=0.25.2,<0.26.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0

# Testing
pytest>=7.4.3,<7.5.0
pytest-asyncio>=0.21.1,<0.22.0





